ai_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
langgraph==0.2.60
langchain-google-genai==2.0.7
langchain-core==0.3.25
python-dotenv==1.0.1
hashlib-compat==1.0.0
